Species data for Kanto
Added in species data for most species within Kanto except for a few
select species that simply didn't make sense for this model.

Added a new positional argument to argparse for which species you wish
to simulate.

Currently only adding in configurable features for egg cycles and sex
ratio of pokemon.

It's getting late and I'm going to bed. Hope this helps some of y'all.